# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation de l'API pour la gestion des utilisateurs et l'authentification.",
        "contact": {
            "email": "sahliabderazak530@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/contactable-users": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Récupérer les utilisateurs contactables",
                "operationId": "getContactableUsers",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs contactables",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/messages": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Envoyer un message",
                "operationId": "sendMessage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "to_user_id",
                                    "content"
                                ],
                                "properties": {
                                    "to_user_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Hello!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message envoyé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/messages/{userId}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Obtenir les messages d'un utilisateur",
                "operationId": "getMessages",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des messages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Message"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/messages/{messageId}/read": {
            "patch": {
                "tags": [
                    "Messages"
                ],
                "summary": "Marquer un message comme lu",
                "operationId": "markAsRead",
                "parameters": [
                    {
                        "name": "messageId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message marqué comme lu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Message non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/messages/read-all/{userId}": {
            "patch": {
                "tags": [
                    "Messages"
                ],
                "summary": "Marquer tous les messages comme lus",
                "operationId": "markAllAsRead",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tous les messages marqués comme lus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/messages/unread-counts": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Obtenir le nombre de messages non lus par utilisateur",
                "operationId": "getUnreadCounts",
                "responses": {
                    "200": {
                        "description": "Nombre de messages non lus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "counts": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/messages/unread-total": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Obtenir le nombre total de messages non lus",
                "operationId": "getUnreadTotal",
                "responses": {
                    "200": {
                        "description": "Nombre total de messages non lus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "count": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Message": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "from_user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "to_user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "content": {
                        "type": "string",
                        "example": "Hello!"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-19T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-19T00:00:00Z"
                    }
                },
                "type": "object"
            },
            "Error": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Non authentifié"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "numTel": {
                        "type": "string",
                        "example": "0123456789"
                    },
                    "adresse": {
                        "type": "string",
                        "example": "123 Main St"
                    },
                    "role": {
                        "type": "string",
                        "example": "admin"
                    },
                    "image": {
                        "type": "string",
                        "example": "profile.jpg"
                    },
                    "archived": {
                        "type": "boolean",
                        "example": false
                    },
                    "nom_societe": {
                        "type": "string",
                        "example": "Example Company"
                    },
                    "active": {
                        "type": "boolean",
                        "example": true
                    },
                    "apropos": {
                        "type": "string",
                        "example": "About the company"
                    },
                    "lien_site_web": {
                        "type": "string",
                        "example": "https://example.com"
                    },
                    "fax": {
                        "type": "string",
                        "example": "0123456789"
                    },
                    "domaine_activite": {
                        "type": "string",
                        "example": "Technology"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}